LayoutTests/http/tests/device-orientation/device-orientation-third-party-iframe-denied.html

<!DOCTYPE html>
<html>
<body>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that third-party iframes cannot request access to device orientation events.");
jsTestIsAsync = true;

// If prompted, allow.
if (window.testRunner)
    testRunner.setShouldAllowDeviceOrientationAndMotionAccess(true);

onmessage = (_e) => {
    e = _e;
    shouldBeEqualToString("e.data", "denied");
    finishJSTest();    
};
</script>
<iframe src="http://localhost:8000/device-orientation/resources/request-orientation-events.html"></iframe>
</body>
</html>
